renamer klient -> client
