+++ settings.gradle.kts
@@ -1,4 +1,4 @@
 rootProject.name = "sosialhjelp-common"
-include("sosialhjelp-common-selftest")
-include("sosialhjelp-common-idporten-klient")
 include("sosialhjelp-common-kotlin-utils")
+include("sosialhjelp-common-selftest")
+include("sosialhjelp-common-idporten-client")
