+++ sosialhjelp-common-idporten-client/src/main/kotlin/no/nav/sosialhjelp/idporten/client/IdPortenProperties.kt
@@ -1,4 +1,4 @@
-package no.nav.sosialhjelp.idporten.klient
+package no.nav.sosialhjelp.idporten.client
 
 data class IdPortenProperties(
     val idPortenTokenUrl: String,
